sửa hàm level

+++ src/TowerDefense/Entity/Enemy/Enemy.java
@@ -32,7 +32,7 @@ public abstract class Enemy extends GameEntity {
     public boolean isDead() {
         return health <= 0;
     }
-    public void takeDamage(int damage) {
+    public void takeDamage(double damage) {
         health -= damage;
     }
